fest(framework): test framework introduction (#53)
