+++ runtime/lib/app-runtime.js
@@ -774,6 +774,42 @@ AppRuntime.prototype.setConfirm = function (appId, intent, slot, options, attrs)
   }).then(() => this.setPickup(true))
 }
 
+AppRuntime.prototype.voiceCommand = function (text, options) {
+  var isTriggered = _.get(options, 'isTriggered', false)
+  var appId = _.get(options, 'appId')
+
+  var skillOption = {
+    device: {
+      linkage: {
+        trigger: isTriggered
+      }
+    }
+  }
+  return new Promise((resolve, reject) => {
+    this.flora.getNlpResult(text, skillOption, function (err, nlp, action) {
+      if (err) {
+        return reject(err)
+      }
+      logger.info('get nlp result for asr', text, nlp, action)
+      resolve([ nlp, action ])
+    })
+  }).then((result) => {
+    var nlp = result[0]
+    var action = result[1]
+    var future = Promise.resolve()
+    if (appId) {
+      /**
+       * retreat self-app into background, then promote the upcoming app
+       * to prevent self being destroy in stack preemption.
+       */
+      future = this.life.setBackgroundById(appId)
+    }
+    return future.then(() => this.onVoiceCommand(text, nlp, action, {
+      carrierId: appId
+    }))
+  })
+}
+
 /**
  *
  * @param {string} appId -
