+++ runtime/lib/descriptor/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+  ActivityDescriptor: require('./activity-descriptor'),
+  ActivityTestDescriptor: require('./activity-test-descriptor'),
+  KeyboardDescriptor: require('./keyboard-descriptor'),
+  LightDescriptor: require('./light-descriptor'),
+  MultimediaDescriptor: require('./multimedia-descriptor'),
+  TtsDescriptor: require('./tts-descriptor'),
+  TurenDescriptor: require('./turen-descriptor'),
+  WormholeDescriptor: require('./wormhole-descriptor')
+}
