+++ test/runtime/ext-app-activity-api.test.js
@@ -4,7 +4,7 @@ var test = require('tape')
 var path = require('path')
 
 var helper = require('../helper')
-var Descriptors = require(`${helper.paths.runtime}/lib/app/activity-descriptor`)
+var Descriptors = require(`${helper.paths.runtime}/lib/descriptor`)
 var extApp = require(`${helper.paths.runtime}/lib/app/ext-app`)
 
 var ActivityDescriptor = Descriptors.ActivityDescriptor
