+++ test/runtime/ext-app.test.js
@@ -5,7 +5,7 @@ var path = require('path')
 var EventEmitter = require('events')
 
 var helper = require('../helper')
-var Descriptors = require(`${helper.paths.runtime}/lib/app/activity-descriptor`)
+var Descriptors = require(`${helper.paths.runtime}/lib/descriptor`)
 var extApp = require(`${helper.paths.runtime}/lib/app/ext-app`)
 
 var ActivityDescriptor = Descriptors.ActivityDescriptor
